# docs/swagger.yaml
definitions:
...
  requests.KubeApplyYamlRequest:
    properties:
      name:
        type: string
      namespace:
        type: string
      yaml:
        type: string
    type: object
...
  requests.KubePVCCreateRequest:
    type: object
  requests.KubePVCResizeRequest:
    properties:
      name:
        type: string
      namespace:
        type: string
      storage:
        description: e.g. "20Gi"
        type: string
    type: object
...
  requests.YamlCreateRequest:
    properties:
      yaml:
        description: YAML 内容
        type: string
    required:
    - yaml
    type: object
  response.Response:
    type: object
...
info:
  contact: {}
  description: 基于Gin+Vue开发的K8s管理平台
  termsOfService: https://gitee.com/jay-kim/k8s_operation
  title: K8s管理平台
  version: "1.0"
paths:
...
  /api/v1/k8s/pvc/apply-yaml:
    put:
      consumes:
      - application/json
      description: 应用修改后的 YAML 配置到 PVC
      parameters:
      - description: YAML内容
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/requests.KubeApplyYamlRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 请求错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 应用 PVC YAML 配置
      tags:
      - K8s PVC 管理
  /api/v1/k8s/pvc/create:
    post:
      consumes:
      - application/json
      description: 创建 PVC（支持指定 StorageClass / AccessModes / 容量等）
      parameters:
      - description: PVC 创建参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/requests.KubePVCCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: 创建 PersistentVolumeClaim
      tags:
      - K8s PVC 管理
  /api/v1/k8s/pvc/create-from-yaml:
    post:
      consumes:
      - application/json
      parameters:
      - description: YAML 创建参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/requests.YamlCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: 从 YAML 创建 PVC
      tags:
      - K8s PVC 管理
  /api/v1/k8s/pvc/delete:
    delete:
      parameters:
      - description: 命名空间
        in: query
        name: namespace
        required: true
        type: string
      - description: PVC 名称
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 删除 PersistentVolumeClaim
      tags:
      - K8s PVC 管理
  /api/v1/k8s/pvc/detail:
    get:
      description: 查询指定命名空间下的 PVC 详情（包括 YAML 内容）
      parameters:
      - description: 命名空间
        in: query
        name: namespace
        required: true
        type: string
      - description: PersistentVolumeClaim 名称
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 资源不存在
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 获取 PersistentVolumeClaim 详情
      tags:
      - K8s PVC 管理
  /api/v1/k8s/pvc/list:
    get:
      description: 支持分页、名称模糊查询（PVC 属于命名空间级资源，需要 namespace 参数）
      parameters:
      - description: 命名空间
        in: query
        name: namespace
        required: true
        type: string
      - description: PVC 名称(模糊匹配)
        in: query
        maxLength: 100
        name: name
        type: string
      - description: 页码(从1开始)
        in: query
        name: page
        required: true
        type: integer
      - description: 每页数量(默认20)
        in: query
        name: limit
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 获取 PVC 列表
      tags:
      - K8s PVC 管理
  /api/v1/k8s/pvc/resize:
    patch:
      consumes:
      - application/json
      description: 将 PVC 的 spec.resources.requests.storage 扩大为指定值（需 StorageClass 允许扩容）
      parameters:
      - description: 扩容参数：namespace/name/storage(如 10Gi)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/requests.KubePVCResizeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
        "403":
          description: StorageClass 不允许扩容
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 资源不存在
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 扩容 PersistentVolumeClaim（仅支持增大 storage）
      tags:
      - K8s PVC 管理
...
swagger: "2.0"